Move pages to own folders - add mainPlan to context
