+++ client/components/mma/switch/SwitchContainer.tsx
@@ -4,9 +4,10 @@ import { Navigate, Outlet, useLocation } from 'react-router';
 import type {
 	MembersDataApiResponse,
 	MembersDataApiUser,
+	PaidSubscriptionPlan,
 	ProductDetail,
 } from '../../../../shared/productResponse';
-import { isProduct } from '../../../../shared/productResponse';
+import { getMainPlan, isProduct } from '../../../../shared/productResponse';
 import { PRODUCT_TYPES } from '../../../../shared/productTypes';
 import {
 	LoadingState,
@@ -30,6 +31,7 @@ export interface SwitchContextInterface {
 	productDetail: ProductDetail;
 	isFromApp: boolean;
 	user?: MembersDataApiUser;
+	mainPlan: PaidSubscriptionPlan;
 }
 
 export const SwitchContext: Context<SwitchContextInterface | {}> =
@@ -121,6 +123,9 @@ const RenderedPage = (props: {
 					productDetail: props.productDetail,
 					isFromApp: props.isFromApp,
 					user: props.user,
+					mainPlan: getMainPlan(
+						props.productDetail.subscription,
+					) as PaidSubscriptionPlan,
 				}}
 			>
 				<Outlet />
