+++ client/components/mma/switch/complete/SwitchSignInImage.tsx
